Export nested behavior events.
