+++ gulpfile.js
@@ -68,32 +68,14 @@ gulp.task('bower:install', ['clean'], function() {
     }));
 });
 
-function getBehaviorPropertiesRecursively(item, name) {
-  var properties = [];
-
-  var behavior = helpers.findBehavior(name)
-  if (behavior) {
-    behavior.properties.forEach(function(prop) {
-      prop.isBehavior = true;
-      prop.behavior = helpers.className(item.is);
-      properties.push(prop);
-    });
-
-    if(behavior.behaviors) {
-      behavior.behaviors.forEach(function(b) {
-        properties = _.union(properties, getBehaviorPropertiesRecursively(item,b));
-      });
-    }
-  }
-
-  return properties;
-}
-
 gulp.task('parse', ['analyze'], function(cb) {
   global.parsed.forEach(function(item) {
     if (!helpers.isBehavior(item) && item.behaviors && item.behaviors.length) {
+
       item.behaviors.forEach(function(name) {
-        item.properties = _.union(item.properties, getBehaviorPropertiesRecursively(item, name));
+        var nestedBehaviors = helpers.getNestedBehaviors(item, name);
+        item.properties = _.union(item.properties, nestedBehaviors.properties);
+        item.events = _.union(item.events, nestedBehaviors.events);
       });
     }
     // Hydrolysis duplicates attributes
